scoring accounts now for lower cuisine-preference

+++ app/FoodHeroTests/Search/Domain/RestaurantBuilder.h
@@ -17,5 +17,5 @@
 
 - (RestaurantBuilder *)withPriceLevel:(NSUInteger)priceLevel;
 
-- (RestaurantBuilder *)withCuisineRelevance:(NSUInteger)cuisineRelevance;
+- (RestaurantBuilder *)withCuisineRelevance:(double)cuisineRelevance;
 @end
